- Add Darwin (OSX) disk status;

+++ util/mystatfs_linux.go
@@ -14,8 +14,8 @@ func GetStatus(stat *MyStatFs, path string) error {
 	}
 
 	// Available blocks * size per block = available space in bytes
-	stat.Avail = unixStatFs.Bavail * uint64(unixStatFs.Bsize)
-	stat.Total = unixStatFs.Blocks * uint64(unixStatFs.Bsize)
+	stat.Avail = uint64(unixStatFs.Bavail) * uint64(unixStatFs.Bsize)
+	stat.Total = uint64(unixStatFs.Blocks) * uint64(unixStatFs.Bsize)
 
 	return nil
 }
